+++ smiggins/templates/js/post.js
@@ -1,7 +1,14 @@
-url = `/api/comments?id=${post_id}&comment=${comment}`;
 type = "comment";
 includeUserLink = true;
 includePostLink = true;
+useOffsetC = true;
+timelines = {
+    random: `/api/comments?id=${post_id}&comment=${comment}&sort=random`,
+    newest: `/api/comments?id=${post_id}&comment=${comment}&sort=newest`,
+    oldest: `/api/comments?id=${post_id}&comment=${comment}&sort=oldest`,
+    liked: `/api/comments?id=${post_id}&comment=${comment}&sort=liked`,
+};
+url = timelines.newest;
 dom("post-text") && dom("post-text").addEventListener("input", postTextInputEvent);
 dom("post") && dom("post").addEventListener("click", function () {
     if (dom("post-text").value) {