+++ smiggins/ts/hashtag.ts
@@ -1,7 +1,14 @@
 declare let hashtag: string;
 
 home = true;
-url = `/api/hashtag/${hashtag}`;
 type = "post";
 includeUserLink = true;
 includePostLink = true;
+timelines = {
+  random: `/api/hashtag/${hashtag}?sort=random`,
+  recent: `/api/hashtag/${hashtag}?sort=recent`,
+  liked: `/api/hashtag/${hashtag}?sort=liked`,
+};
+
+url = timelines.random;
+useOffsetC = true;